@@ -28,12 +28,12 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
+        "php": ">=8.0",
         "ext-dom": "*",
         "hans-thomas/common": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "phpunit/phpunit": "~9.0"
     },
     "autoload": {
         "psr-4": {
