@@ -49,7 +49,7 @@ class NavPointList implements Iterator
      * @link http://php.net/manual/en/iterator.current.php
      * @return NavPoint
      */
-    public function current()
+    public function current(): NavPoint
     {
         return current($this->navPoints);
     }
@@ -60,7 +60,7 @@ class NavPointList implements Iterator
      * @link http://php.net/manual/en/iterator.next.php
      * @return void Any returned value is ignored.
      */
-    public function next()
+    public function next(): void
     {
         next($this->navPoints);
     }
@@ -71,7 +71,7 @@ class NavPointList implements Iterator
      * @link http://php.net/manual/en/iterator.key.php
      * @return mixed scalar on success, or null on failure.
      */
-    public function key()
+    public function key(): mixed
     {
         return key($this->navPoints);
     }
@@ -83,7 +83,7 @@ class NavPointList implements Iterator
      * @return boolean The return value will be casted to boolean and then evaluated.
      * Returns true on success or false on failure.
      */
-    public function valid()
+    public function valid(): bool
     {
         return (bool)current($this->navPoints);
     }
@@ -94,7 +94,7 @@ class NavPointList implements Iterator
      * @link http://php.net/manual/en/iterator.rewind.php
      * @return void Any returned value is ignored.
      */
-    public function rewind()
+    public function rewind(): void
     {
         reset($this->navPoints);
     }
