@@ -15,7 +15,7 @@ use Iterator;
  */
 class Spine implements Iterator, Countable, ArrayAccess
 {
-    /** @var string */
+    /** @var Item */
     private $tocItem;
     /** @var array|Item[] The ordered list of all Items in this Spine. */
     private $items = [];
@@ -56,7 +56,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.current.php
      * @return Item
      */
-    public function current()
+    public function current(): Item
     {
         return current($this->items);
     }
@@ -66,7 +66,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.next.php
      * @return void Any returned value is ignored.
      */
-    public function next()
+    public function next(): void
     {
         next($this->items);
     }
@@ -77,7 +77,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.key.php
      * @return int on success, or null on failure.
      */
-    public function key()
+    public function key(): int
     {
         return key($this->items);
     }
@@ -88,7 +88,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.valid.php
      * @return boolean true on success or false on failure.
      */
-    public function valid()
+    public function valid(): bool
     {
         return (bool)current($this->items);
     }
@@ -99,7 +99,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.rewind.php
      * @return void Any returned value is ignored.
      */
-    public function rewind()
+    public function rewind(): void
     {
         reset($this->items);
     }
@@ -130,7 +130,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @link https://php.net/manual/en/countable.count.php
      * @return int The number of Items contained in this Spine.
      */
-    public function count()
+    public function count(): int
     {
         return count($this->items);
     }
@@ -141,7 +141,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @param int $offset An offset to check for.
      * @return boolean true on success or false on failure.
      */
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return isset($this->items[$offset]);
     }
@@ -164,7 +164,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @param mixed $value The value to set.
      * @throws NotSupportedException
      */
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         throw new NotSupportedException("Only reading array access is supported!");
     }
@@ -175,7 +175,7 @@ class Spine implements Iterator, Countable, ArrayAccess
      * @param mixed $offset The offset to unset.
      * @throws NotSupportedException
      */
-    public function offsetUnset($offset)
+    public function offsetUnset($offset): void
     {
         throw new NotSupportedException("Only reading array access is supported!");
     }
