@@ -23,18 +23,18 @@ class Manifest implements Iterator, Countable, ArrayAccess
      *
      * @param string $id The identifier of the new item.
      * @param string $href The relative path of the referenced file in the EPUB.
-     * @param resource $handle A file handle to the referenced file in the EPUB.
+     * @param callable $callable A callable to get data from the referenced file in the EPUB.
      * @param int $size The size of the referenced file in the EPUB.
      * @param string|null $mediaType
      * @return Item The newly created Item.
      * @throws Exception If $id is already taken.
      */
-    public function createItem($id, $href, $handle, $size, $mediaType = null)
+    public function createItem($id, $href, $callable, $size, $mediaType = null)
     {
         if (isset($this->items[$id])) {
             throw new Exception("Item with ID $id already exists!");
         }
-        $item = new Item($id, $href, $handle, $size, $mediaType);
+        $item = new Item($id, $href, $callable, $size, $mediaType);
         $this->items[$id] = $item;
 
         return $item;
@@ -46,7 +46,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.current.php
      * @return Item
      */
-    public function current()
+    public function current(): Item
     {
         return current($this->items);
     }
@@ -56,7 +56,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.next.php
      * @return void Any returned value is ignored.
      */
-    public function next()
+    public function next(): void
     {
         next($this->items);
     }
@@ -67,7 +67,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.key.php
      * @return string on success, or null on failure.
      */
-    public function key()
+    public function key(): string
     {
         return key($this->items);
     }
@@ -78,7 +78,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.valid.php
      * @return boolean true on success or false on failure.
      */
-    public function valid()
+    public function valid(): bool
     {
         return (bool)current($this->items);
     }
@@ -89,7 +89,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @link http://php.net/manual/en/iterator.rewind.php
      * @return void Any returned value is ignored.
      */
-    public function rewind()
+    public function rewind(): void
     {
         reset($this->items);
     }
@@ -120,7 +120,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @link https://php.net/manual/en/countable.count.php
      * @return int The number of Items contained in this Manifest.
      */
-    public function count()
+    public function count(): int
     {
         return count($this->items);
     }
@@ -131,7 +131,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @param int $offset An offset to check for.
      * @return boolean true on success or false on failure.
      */
-    public function offsetExists($offset)
+    public function offsetExists($offset): bool
     {
         return isset($this->items[$offset]);
     }
@@ -142,7 +142,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @param int $offset The offset to retrieve.
      * @return Item
      */
-    public function offsetGet($offset)
+    public function offsetGet($offset): Item
     {
         return $this->items[$offset];
     }
@@ -154,7 +154,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @param mixed $value The value to set.
      * @throws NotSupportedException
      */
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         throw new NotSupportedException("Only reading array access is supported!");
     }
@@ -165,7 +165,7 @@ class Manifest implements Iterator, Countable, ArrayAccess
      * @param mixed $offset The offset to unset.
      * @throws NotSupportedException
      */
-    public function offsetUnset($offset)
+    public function offsetUnset($offset): void
     {
         throw new NotSupportedException("Only reading array access is supported!");
     }
